services/ai-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
diffusers==0.31.0
transformers==4.46.0
torch==2.5.0
accelerate==1.1.0
Pillow==11.0.0
pydantic==2.9.0
python-multipart==0.0.12
httpx==0.27.2
